Creating Users,Auth API

+++ server/controllers/UserController.js
@@ -3,7 +3,9 @@ import argon2 from "argon2";
 
 export const getUsers = async (req, res) => {
   try {
-    const users = await User.find({});
+    const users = await User.find({}).select(
+      "name email newMessage status password"
+    );
     res.status(200).json(users);
   } catch (error) {
     res.status(400).json({ msg: error.message });
@@ -12,7 +14,9 @@ export const getUsers = async (req, res) => {
 
 export const getUserById = async (req, res) => {
   try {
-    const user = await User.findOne({ _id: req.params.id });
+    const user = await User.findOne({ _id: req.params.id }).select(
+      "name email newMessage status"
+    );
     res.status(200).json(user);
   } catch (error) {
     res.status(400).json({ msg: error.message });
@@ -22,10 +26,11 @@ export const getUserById = async (req, res) => {
 export const createUsers = async (req, res) => {
   try {
     const { name, email, password, img } = req.body;
+    const hashPassword = await argon2.hash(password);
     await User.create({
       name: name,
       email: email,
-      password: password,
+      password: hashPassword,
       img: img,
     });
     console.log(req.body);
@@ -57,6 +62,11 @@ export const updateUser = async (req, res) => {
 
 export const deleteUser = async (req, res) => {
   try {
+    const user = await User.findOne({ _id: req.params.id });
+    console.log(user);
+    if (!user) return res.status(404).json({ msg: "User Not Found" });
+    if (user.email !== req.session.email)
+      return res.status(403).json({ msg: "Akses Terlarang" });
     await User.deleteOne({
       _id: req.params.id,
     });
